Fix #83: Move `JSPlatform` back to `scalajscrossproject`
