+++ sbt-scalajs-crossproject/src/main/scala/scalajscrossproject/JSPlatform.scala
@@ -1,11 +1,8 @@
-/* It is totally evil to introduce a class in the package of sbt-scalajs, but that is
- * the only way we can be forward source compatible with Scala.js 1.x.
- */
-package org.scalajs.sbtplugin
+package scalajscrossproject
 
 import sbtcrossproject._
-
 import sbt._
+import org.scalajs.sbtplugin._
 
 case object JSPlatform extends Platform {
   def identifier: String                = "js"
